requirements.txt
arrow==0.12.1
boto3==1.9.66
botocore==1.12.66
certifi==2018.11.29
chardet==3.0.4
Click==7.0
docopt==0.6.2
docutils==0.14
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kafka==1.3.5
MarkupSafe==1.1.0
mercury-toolkit==0.7.7
python-dateutil==2.7.5
PyYAML==3.13
raven==6.9.0
redis==3.0.1
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
snap-micro==0.9.91
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.33.9
urllib3==1.24.1
Werkzeug==0.14.1
